remove Granule in favor of DispProduct
print
